Убрана проверка на экзмепляр класса, где это возможно(в Human.java в особенности). Добавлен ещё один статус в Status.java, чтобы различать функции методов descend() и moveToHuman(). Кроме этого, все классы были распределены по пакетам иначе(Всех людей в Entities, все локации в Locations, перечисления в Enums и интерфейсы в Interfaces)
