--- Lab3/src/Common/ObjectStatus.java
@@ -1,6 +1,0 @@
-package Common;
-
-public enum ObjectStatus {
-    CLOSED,
-    OPENED
-}
