+++ Lab3/src/Enums/ActiveStatus.java
@@ -1,0 +1,6 @@
+package Enums;
+
+public enum ActiveStatus {
+    SMOKING,
+    NO_SMOKING
+}
