+++ Lab3/src/Enums/FurnaceStatus.java
@@ -1,0 +1,16 @@
+package Enums;
+
+public enum FurnaceStatus {
+    COLD("остывшая"),
+    HOT("горячая");
+    private final String furnaceStatus;
+
+    FurnaceStatus(String furnaceStatus) {
+        this.furnaceStatus = furnaceStatus;
+    }
+
+    @Override
+    public String toString() {
+        return furnaceStatus;
+    }
+}
