+++ Lab3/src/Enums/ObjectStatus.java
@@ -1,0 +1,6 @@
+package Enums;
+
+public enum ObjectStatus {
+    CLOSED,
+    OPENED
+}
