+++ Lab3/src/Enums/Prepositions.java
@@ -1,0 +1,18 @@
+package Enums;
+
+public enum Prepositions {
+    IN("в"),
+    TO("к"),
+    ON("на"),
+    FOR("за");
+    private final String prep;
+
+    Prepositions(String prep) {
+        this.prep = prep;
+    }
+
+    @Override
+    public String toString() {
+        return prep;
+    }
+}
