+++ Lab3/src/Enums/Status.java
@@ -1,0 +1,7 @@
+package Enums;
+
+public enum Status {
+    ACTIVE,
+    STANDING,
+    SITTING
+}
