+++ Lab3/src/Enums/Traits.java
@@ -1,0 +1,18 @@
+package Enums;
+
+public enum Traits {
+    LOUD("громко"),
+    ABRASIVELY("резко");
+
+
+    private final String trait;
+
+    Traits(String trait) {
+        this.trait = trait;
+    }
+
+    @Override
+    public String toString() {
+        return trait;
+    }
+}
