+++ Lab3/src/Enums/TraitsOfLogs.java
@@ -1,0 +1,6 @@
+package Enums;
+
+public enum TraitsOfLogs {
+    NON_BURNING,
+    BURNING
+}
