--- Lab3/src/Humans/ActiveStatus.java
@@ -1,6 +1,0 @@
-package Humans;
-
-public enum ActiveStatus {
-    SMOKING,
-    NO_SMOKING
-}
