--- Lab3/src/Humans/DescendAble.java
@@ -1,5 +1,0 @@
-package Humans;
-
-interface DescendAble {
-    void descend(Human human);
-}
