--- Lab3/src/Humans/GrabAble.java
@@ -1,9 +1,0 @@
-package Humans;
-
-import Location.Storeroom;
-
-public interface GrabAble {
-    void grabLogs(Storeroom storeroom);
-
-    void grabPot(Storeroom storeroom);
-}
