--- Lab3/src/Humans/LightUpAble.java
@@ -1,5 +1,0 @@
-package Humans;
-
-interface LightUpAble {
-    boolean LightUp();
-}
