--- Lab3/src/Humans/MoveAble.java
@@ -1,6 +1,0 @@
-package Humans;
-
-
-interface MoveAble {
-    void moveToHuman(Human human);
-}
