--- Lab3/src/Humans/PutAble.java
@@ -1,10 +1,0 @@
-package Humans;
-
-import Location.Kitchen;
-import Location.Lobby;
-
-public interface PutAble {
-    void putLogs(Kitchen kitchen);
-
-    void putPot(Lobby lobby);
-}
