--- Lab3/src/Humans/Status.java
@@ -1,6 +1,0 @@
-package Humans;
-
-public enum Status {
-    ACTIVE,
-    STATIC
-}
