--- Lab3/src/Humans/Traits.java
@@ -1,18 +1,0 @@
-package Humans;
-
-enum Traits {
-    LOUD("громко"),
-    ABRASIVELY("резко");
-
-
-    private final String trait;
-
-    Traits(String trait) {
-        this.trait = trait;
-    }
-
-    @Override
-    public String toString() {
-        return trait;
-    }
-}
