+++ Lab3/src/Interfaces/CloseAble.java
@@ -1,0 +1,6 @@
+package Interfaces;
+
+public interface CloseAble
+{
+    boolean isClosed();
+}
