+++ Lab3/src/Interfaces/DescendAble.java
@@ -1,0 +1,7 @@
+package Interfaces;
+
+import Entities.Human;
+
+public interface DescendAble {
+    void descend(Human human);
+}
