+++ Lab3/src/Interfaces/EnterLocationAble.java
@@ -1,0 +1,6 @@
+package Interfaces;
+
+public interface EnterLocationAble
+{
+    void enterLocation(ILocation location);
+}
