+++ Lab3/src/Interfaces/ILocation.java
@@ -1,0 +1,6 @@
+package Interfaces;
+
+public interface ILocation {
+    void describe();
+
+}
