+++ Lab3/src/Interfaces/LightUpAble.java
@@ -1,0 +1,5 @@
+package Interfaces;
+
+public interface LightUpAble {
+    boolean LightUp();
+}
