+++ Lab3/src/Interfaces/MoveAble.java
@@ -1,0 +1,8 @@
+package Interfaces;
+
+
+import Entities.Human;
+
+public interface MoveAble {
+    void moveToHuman(Human human);
+}
