--- Lab3/src/Location/FurnaceStatus.java
@@ -1,16 +1,0 @@
-package Location;
-
-public enum FurnaceStatus {
-    COLD("остывшая"),
-    HOT("горячая");
-    private final String furnaceStatus;
-
-    FurnaceStatus(String furnaceStatus) {
-        this.furnaceStatus = furnaceStatus;
-    }
-
-    @Override
-    public String toString() {
-        return furnaceStatus;
-    }
-}
