--- Lab3/src/Location/ILocation.java
@@ -1,6 +1,0 @@
-package Location;
-
-public interface ILocation {
-    void describe();
-
-}
