--- Lab3/src/Location/Room.java
@@ -1,6 +1,0 @@
-package Location;
-
-public abstract class Room implements ILocation {
-
-
-}
