--- Lab3/src/Location/TraitsOfClock.java
@@ -1,16 +1,0 @@
-package Location;
-
-public enum TraitsOfClock {
-    WORKING("ещё работавшие"),
-    STOPPED("остановившиеся");
-    private final String traitOfClock;
-
-    TraitsOfClock(String traitOfClock) {
-        this.traitOfClock = traitOfClock;
-    }
-
-    @Override
-    public String toString() {
-        return traitOfClock;
-    }
-}
