--- Lab3/src/Location/TraitsOfFlowers.java
@@ -1,16 +1,0 @@
-package Location;
-
-public enum TraitsOfFlowers {
-    NOTFADED("ещё свежие"),
-    FADED("увядшие");
-    private final String traitOfFlowers;
-
-    TraitsOfFlowers(String traitOfObjects) {
-        this.traitOfFlowers = traitOfObjects;
-    }
-
-    @Override
-    public String toString() {
-        return traitOfFlowers;
-    }
-}
