--- Lab3/src/Location/TraitsOfLogs.java
@@ -1,6 +1,0 @@
-package Location;
-
-public enum TraitsOfLogs {
-    NON_BURNING,
-    BURNING
-}
