--- Lab3/src/Location/Weather.java
@@ -1,15 +1,0 @@
-package Location;
-
-public enum Weather {
-    RAINY("шёл дождь");
-    private final String weather;
-
-    Weather(String weather) {
-        this.weather = weather;
-    }
-
-    @Override
-    public String toString() {
-        return weather;
-    }
-}
