+++ Lab3/src/Locations/Room.java
@@ -1,0 +1,6 @@
+package Locations;
+import Interfaces.ILocation;
+public abstract class Room implements ILocation {
+
+
+}
