+++ LabTwo/LabTwo.iml
@@ -7,14 +7,9 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module-library">
-      <library>
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../Pokemon.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
+    <orderEntry type="library" name="LabTwo1" level="project" />
+    <orderEntry type="library" name="LabTwo" level="project" />
+    <orderEntry type="library" name="Pokemon" level="project" />
+    <orderEntry type="library" name="LabTwo2" level="project" />
   </component>
 </module>
